refactor(input): fix name DSInputNumber and DSInputPassword
