+++ libs/ui/src/components/RecordEdition/EditRecordContent/uiElements/StandardField/StandardFieldValue/DSDatePickerWrapper.test.tsx
@@ -2,7 +2,7 @@
 // This file is released under LGPL V3
 // License text available at https://www.gnu.org/licenses/lgpl-3.0.txt
 import {render, screen} from '_ui/_tests/testUtils';
-import {DSDatePickerWrapper} from './DSDatePickerWrapperWrapper';
+import {DSDatePickerWrapper} from './DSDatePickerWrapper';
 import {FieldScope} from '../../../_types';
 import {
     IStandardFieldReducerState,
